+++ src/lib.rs
@@ -57,7 +57,7 @@ mod syntax;
 mod text;
 
 pub use completing::Completion;
-pub use editor::{Editor, EditorOutput, EditorRequest, Marks};
+pub use editor::{Editor, EditorOutput, EditorRequest, Marks, Underline};
 pub use indenting::Indent;
 pub use place::{TextPoint, Word};
 pub use style::{EditorStyle, SyntaxTheme, TokenLook};